FluentValidators and OneOf implementation (#2)
* Added item validator

* Introduced oneOf

* OneOf refactoring

* Added get all lists and items endpoints

---------

+++ TodoList/Services/ListService.cs
@@ -1,8 +1,12 @@
 ﻿using AutoMapper;
+using FluentValidation;
 using Microsoft.EntityFrameworkCore;
+using OneOf;
+using OneOf.Types;
 using TodoList.ModelExtensions;
 using TodoList.Models;
 using TodoList.Repository;
+using TodoList.Validators;
 using TodoList.ViewModels;
 
 namespace TodoList.Services
@@ -11,17 +15,44 @@ namespace TodoList.Services
     {
         private readonly TodoContext _context;
         private readonly IMapper _mapper;
+        private readonly IValidator<ItemsList> _validator;
 
-        public ListService(TodoContext context, IMapper mapper)
+        public ListService(TodoContext context, IMapper mapper, IValidator<ItemsList> validator)
         {
             _context = context;
             _mapper = mapper;
+            _validator = validator;
+        }
+        public OneOf<ListViewModel, NotFound> GetById(Guid listId)
+        {
+            var list =  _context.ItemsLists
+                .Include(list => list.Items)
+                .FirstOrDefault(list => list.Id == listId);
+
+            if (list == null)
+            {
+                return new NotFound();
+            }
+
+            return _mapper.Map<ListViewModel>(list);
         }
 
-        public ListViewModel Create(CreateListViewModel viewModel)
+        public List<ListViewModel> GetAllLists()
+        {
+            var lists = _context.ItemsLists.ToArray();
+            return _mapper.Map<ItemsList[], List<ListViewModel>>(lists);
+        }
+
+        public OneOf<ListViewModel, ValidationFailed> Create(CreateListViewModel viewModel)
         {
             var list = _mapper.Map<ItemsList>(viewModel);
             list.SetIdsAndDates();
+
+            var result = _validator.Validate(list);
+            if (!result.IsValid)
+            {
+                return new ValidationFailed(result.Errors);
+            }
 
             var entity  = _context.ItemsLists.Add(list);
             _context.SaveChanges();
@@ -29,30 +60,37 @@ namespace TodoList.Services
             return _mapper.Map<ListViewModel>(entity.Entity);
         }
 
-        public void Delete(Guid listId)
-        {
-            var list = new ItemsList { Id= listId };
-            _context.ItemsLists.Attach(list);
-            _context.ItemsLists.Remove(list);
-            _context.SaveChanges();
-        }
-
-        public ListViewModel GetById(Guid listId)
-        {
-            var list =  _context.ItemsLists
-                .Include(list => list.Items)
-                .FirstOrDefault(list => list.Id == listId)!;
-
-            return _mapper.Map<ListViewModel>(list);
-        }
-
-        public ListViewModel Update(Guid listId, UpdateListViewModel list)
+        public OneOf<ListViewModel, ValidationFailed, NotFound> Update(Guid listId, UpdateListViewModel list)
         {
             var entity = _context.ItemsLists.FirstOrDefault(list => list.Id == listId);
+            if (entity == null)
+            {
+                return new NotFound();
+            }
             entity.Update(list);
-            _context.SaveChanges();
 
+            var result = _validator.Validate(entity);
+            if (!result.IsValid)
+            {
+                return new ValidationFailed(result.Errors);
+            }
+
+            _context.SaveChanges();
             return _mapper.Map<ListViewModel>(entity);
+        }
+
+        public OneOf<Success, NotFound> Delete(Guid listId)
+        {
+            var list = _context.ItemsLists.FirstOrDefault(list => list.Id == listId);
+
+            if (list == null)
+            {
+                return new NotFound();
+            }
+
+            _context.ItemsLists.Remove(list);
+            _context.SaveChanges();
+            return new Success();
         }
     }
 }
